+++ lib/app/presentation/app_root.dart
@@ -1,10 +1,17 @@
+import 'dart:async';
+
+import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 import 'package:provider/provider.dart';
 import 'package:zachranobed/common/domain/model/app_terms_status.dart';
+import 'package:zachranobed/common/domain/model/user_data.dart';
 import 'package:zachranobed/common/domain/repository/delivery_repository.dart';
 import 'package:zachranobed/common/domain/usecase/get_app_terms_status_usecase.dart';
 import 'package:zachranobed/common/domain/usecase/get_user_data_usecase.dart';
+import 'package:zachranobed/common/domain/usecase/observe_user_data_usecase.dart';
+import 'package:zachranobed/common/domain/usecase/remove_onboarding_for_ui_changes_flag_usecase.dart';
+import 'package:zachranobed/common/domain/usecase/should_show_onboarding_for_ui_changes_usecase.dart';
 import 'package:zachranobed/common/domain/utils/platform_utils.dart';
 import 'package:zachranobed/common/presentation/notifiers/delivery_notifier.dart';
 import 'package:zachranobed/common/presentation/notifiers/user_notifier.dart';
@@ -12,7 +19,6 @@ import 'package:zachranobed/common/presentation/router/app_router.dart';
 import 'package:zachranobed/common/presentation/router/app_router.gr.dart';
 import 'package:zachranobed/common/presentation/utils/lifecycle_watcher.dart';
 import 'package:zachranobed/common/presentation/utils/ui_colors.dart';
-import 'package:zachranobed/common/presentation/utils/ui_constants.dart';
 import 'package:zachranobed/common/presentation/utils/ui_text_styles.dart';
 import 'package:zachranobed/features/forceupdate/domain/usecase/check_if_upgrade_app_should_be_shown_usecase.dart';
 import 'package:zachranobed/features/offline/presentation/connectivity_wrapper.dart';
@@ -28,14 +34,31 @@ class AppRoot extends StatefulWidget {
 class _AppRootState extends State<AppRoot> with LifecycleWatcher {
   final _appRouter = GetIt.I<AppRouter>();
   final _getUserData = GetIt.I<GetUserDataUseCase>();
+  final _observeUserData = GetIt.I<ObserveUserDataUseCase>();
   final _checkIfUpgradeAppShouldBeShown = GetIt.I<CheckIfUpgradeAppShouldBeShownUseCase>();
   final _getAppTermsStatus = GetIt.I<GetAppTermsStatusUseCase>();
+  final _shouldShowOnboardingForUiChanges = GetIt.I<ShouldShowOnboardingForUiChangesUseCase>();
+  final _removeOnboardingForUiChangesFlag = GetIt.I<RemoveOnboardingForUiChangesFlagUseCase>();
+
+  StreamSubscription<void>? _userDataSubscription;
 
   @override
   void initState() {
     super.initState();
 
+    _userDataSubscription = _observeUserData.invoke().listen((user) {
+      if (user != null) {
+        _applicationStartCheckForUser(user);
+      }
+    });
+
     _applicationStartCheck();
+  }
+
+  @override
+  void dispose() {
+    _userDataSubscription?.cancel();
+    super.dispose();
   }
 
   @override
@@ -43,25 +66,35 @@ class _AppRootState extends State<AppRoot> with LifecycleWatcher {
     _applicationStartCheck();
   }
 
+  /// Performs the initial checks after the application starts.
+  /// 1. Check if the app should be updated.
+  /// 2. Perform for user-related checks, see [_applicationStartCheckForUser].
   void _applicationStartCheck() async {
     final shouldShow = await _checkIfUpgradeAppShouldBeShown.invoke();
     if (shouldShow) {
       _appRouter.replace(const ForceUpdateRoute());
     }
 
-    _checkAppTerms();
+    final user = await _getUserData.invoke();
+    if (user != null) {
+      _applicationStartCheckForUser(user);
+    }
   }
 
-  void _checkAppTerms() async {
-    final user = await _getUserData.invoke();
-    if (user == null) {
-      // User is not logged in, do not check app terms
-      return;
-    }
-
+  /// Performs the initial checks after the application starts.
+  /// 1. Check if the app terms are accepted.
+  /// 2. Check if the onboarding for UI changes should be shown.
+  void _applicationStartCheckForUser(UserData user) async {
     final status = await _getAppTermsStatus.invoke(user);
     if (status != AppTermsStatus.accepted) {
       _appRouter.replace(AppTermsRoute(hasNoAcceptedVersion: status == AppTermsStatus.notAccepted));
+      return;
+    }
+
+    final shouldShowOnboarding = await _shouldShowOnboardingForUiChanges.invoke(user.entityId);
+    if (shouldShowOnboarding) {
+      await _removeOnboardingForUiChangesFlag.invoke(user.entityId);
+      _appRouter.push(const WhatsNewRoute());
     }
   }
 
@@ -79,34 +112,37 @@ class _AppRootState extends State<AppRoot> with LifecycleWatcher {
           // It is not possible to use AppLocalizations here, because it is not
           // available yet. Therefore, the title is hardcoded.
           title: "Zachraň oběd",
-          routerConfig: _appRouter.config(),
+          routerConfig: _appRouter.config(
+            navigatorObservers: () => [
+              FirebaseAnalyticsObserver(analytics: FirebaseAnalytics.instance),
+            ],
+          ),
           localizationsDelegates: AppLocalizations.localizationsDelegates,
           supportedLocales: AppLocalizations.supportedLocales,
           locale: const Locale('cs'),
           theme: ThemeData(
             useMaterial3: true,
             colorScheme: ColorScheme.fromSwatch().copyWith(
-              primary: ZOColors.primary,
-              onPrimary: ZOColors.onPrimary,
-              secondary: ZOColors.primaryLight,
-              onSecondary: ZOColors.onPrimaryLight,
-              surface: Colors.white,
-              surfaceTint: Colors.white,
-              primaryContainer: ZOColors.secondary,
+              primary: UiColors.light.primary,
+              onPrimary: UiColors.light.textPrimaryInverse,
+              secondary: UiColors.light.primary,
+              onSecondary: UiColors.light.textPrimaryInverse,
+              surface: UiColors.light.surfaceWhite,
+              surfaceTint: UiColors.light.surfaceWhite,
             ),
             textTheme: UiTextStyles.getTextTheme(),
-            scaffoldBackgroundColor: Colors.white,
+            scaffoldBackgroundColor: UiColors.light.surfaceWhite,
             visualDensity: VisualDensity.standard,
-            highlightColor: Colors.transparent,
-            splashColor: Colors.transparent,
+            highlightColor: UiColors.light.transparent,
+            splashColor: UiColors.light.transparent,
             appBarTheme: AppBarTheme(
               // Center title only for iOS platform
               centerTitle: RunningPlatform.isIOS(),
               scrolledUnderElevation: 0,
             ),
-            bottomSheetTheme: const BottomSheetThemeData(
+            bottomSheetTheme: BottomSheetThemeData(
               showDragHandle: true,
-              dragHandleColor: ZOColors.outline,
+              dragHandleColor: UiColors.light.textSecondary,
             ),
             materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
             extensions: [