+++ lib/app/presentation/home_screen.dart
@@ -6,6 +6,7 @@ import 'package:get_it/get_it.dart';
 import 'package:provider/provider.dart';
 import 'package:zachranobed/app/presentation/overview_screen.dart';
 import 'package:zachranobed/common/data/service/auth_service.dart';
+import 'package:zachranobed/common/domain/usecase/notify_user_data_changed_usecase.dart';
 import 'package:zachranobed/common/presentation/notifiers/user_notifier.dart';
 import 'package:zachranobed/common/presentation/router/app_router.gr.dart';
 import 'package:zachranobed/common/presentation/utils/build_context_extensions.dart';
@@ -15,9 +16,9 @@ import 'package:zachranobed/common/presentation/utils/lifecycle_watcher.dart';
 import 'package:zachranobed/common/presentation/utils/ui_constants.dart';
 import 'package:zachranobed/common/presentation/widget/button/ui_primary_button.dart';
 import 'package:zachranobed/common/presentation/widget/navigation/ui_nav_bar.dart';
-import 'package:zachranobed/common/presentation/widget/screen_scaffold.dart';
-import 'package:zachranobed/common/presentation/widget/ui_icon.dart';
-import 'package:zachranobed/common/presentation/widget/ui_navigation_drawer_item.dart';
+import 'package:zachranobed/common/presentation/widget/layout/screen_scaffold.dart';
+import 'package:zachranobed/common/presentation/widget/graphics/ui_icon.dart';
+import 'package:zachranobed/common/presentation/widget/navigation/ui_navigation_drawer_item.dart';
 import 'package:zachranobed/features/food/presentation/screens/history_screen.dart';
 import 'package:zachranobed/features/notifications/domain/usecase/has_any_unread_notifications_use_case.dart';
 import 'package:zachranobed/features/notifications/domain/usecase/update_notifications_token_use_case.dart';
@@ -33,6 +34,7 @@ class HomeScreen extends StatefulWidget {
 
 class _HomeScreenState extends State<HomeScreen> with LifecycleWatcher, SingleTickerProviderStateMixin {
   final _updateNotificationsTokenUseCase = GetIt.I<UpdateNotificationsTokenUseCase>();
+  final _notifyUserDataChanged = GetIt.I<NotifyUserDataChangedUseCase>();
 
   /// The data for the tabs.
   final _tabs = [
@@ -204,6 +206,7 @@ class _HomeScreenState extends State<HomeScreen> with LifecycleWatcher, SingleTi
                   onPressed: () async {
                     final entityId = HelperService.getCurrentUser(context)?.entityId;
                     await authService.signOut(entityId);
+                    _notifyUserDataChanged.invoke(null);
                     if (mounted) {
                       context.router.replaceAll([const LoginRoute()]);
                     }