+++ lib/common/presentation/widget/donation/ui_donation_status_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:zachranobed/common/presentation/utils/build_context_extensions.dart';
-import 'package:zachranobed/common/presentation/widget/ui_card.dart';
+import 'package:zachranobed/common/presentation/widget/card/ui_card.dart';
 
 /// A card widget for displaying donation status information with optional progress tracking, status messages,
 /// and actions.