+++ lib/features/food/presentation/screens/offer_food_overview_screen.dart
@@ -11,18 +11,18 @@ import 'package:zachranobed/common/presentation/widget/button/ui_button_size.dar
 import 'package:zachranobed/common/presentation/widget/button/ui_icon_button.dart';
 import 'package:zachranobed/common/presentation/widget/button/ui_primary_button.dart';
 import 'package:zachranobed/common/presentation/widget/button/ui_text_button.dart';
-import 'package:zachranobed/common/presentation/widget/other/adaptive_content.dart';
-import 'package:zachranobed/common/presentation/widget/other/section_header.dart';
+import 'package:zachranobed/common/presentation/widget/layout/adaptive_content.dart';
+import 'package:zachranobed/common/presentation/widget/layout/section_header.dart';
 import 'package:zachranobed/common/presentation/widget/page/info_page.dart';
-import 'package:zachranobed/common/presentation/widget/screen_scaffold.dart';
-import 'package:zachranobed/common/presentation/widget/snackbar/temporary_snackbar.dart';
-import 'package:zachranobed/common/presentation/widget/ui_app_bar.dart';
-import 'package:zachranobed/common/presentation/widget/ui_checkbox.dart';
-import 'package:zachranobed/common/presentation/widget/ui_dialog.dart';
-import 'package:zachranobed/common/presentation/widget/ui_gradient_icon.dart';
-import 'package:zachranobed/common/presentation/widget/ui_icon.dart';
-import 'package:zachranobed/common/presentation/widget/ui_list_tile.dart';
-import 'package:zachranobed/common/presentation/widget/ui_notification_tile.dart';
+import 'package:zachranobed/common/presentation/widget/layout/screen_scaffold.dart';
+import 'package:zachranobed/common/presentation/widget/overlay/ui_temporary_snackbar.dart';
+import 'package:zachranobed/common/presentation/widget/navigation/ui_app_bar.dart';
+import 'package:zachranobed/common/presentation/widget/form/ui_checkbox.dart';
+import 'package:zachranobed/common/presentation/widget/overlay/ui_dialog.dart';
+import 'package:zachranobed/common/presentation/widget/graphics/ui_gradient_icon.dart';
+import 'package:zachranobed/common/presentation/widget/graphics/ui_icon.dart';
+import 'package:zachranobed/common/presentation/widget/card/ui_list_tile.dart';
+import 'package:zachranobed/common/presentation/widget/card/ui_notification_tile.dart';
 import 'package:zachranobed/features/food/domain/model/food_box_type.dart';
 import 'package:zachranobed/features/food/domain/model/food_info.dart';
 import 'package:zachranobed/features/food/domain/usecase/create_food_offer_use_case.dart';
@@ -270,12 +270,7 @@ class _OfferFoodOverviewScreenState extends State<OfferFoodOverviewScreen> {
     } else {
       if (mounted) {
         context.router.maybePop();
-        ScaffoldMessenger.of(context).showSnackBar(
-          ZOTemporarySnackBar(
-            backgroundColor: Colors.red,
-            message: context.l10n.boxCountError,
-          ),
-        );
+        UiTemporarySnackBar.showError(context, message: context.l10n.boxCountError);
       }
     }
   }