+++ lib/features/offline/presentation/offline_screen.dart
@@ -6,7 +6,7 @@ import 'package:zachranobed/common/presentation/utils/image_assets.dart';
 import 'package:zachranobed/common/presentation/widget/button/ui_button_size.dart';
 import 'package:zachranobed/common/presentation/widget/button/ui_primary_button.dart';
 import 'package:zachranobed/common/presentation/widget/page/info_page.dart';
-import 'package:zachranobed/common/presentation/widget/screen_scaffold.dart';
+import 'package:zachranobed/common/presentation/widget/layout/screen_scaffold.dart';
 
 /// A screen displayed when the application detects that there is no internet connectivity.
 /// The button to close an application is displayed only on Android, as per iOS guidelines it is discouraged to have