+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This is a changelog for **ZachranObed** application.
 ### Added
 - **ZOB-381** Add new UI components and color definitions from new design system.
 - **ZOB-393** Add new delivery details screen.
+- **ZOB-404** Add new "What's new" screen.
+- **ZOB-405** Add support for allergen sub categories.
 
 ### Fixed
 
@@ -30,6 +32,7 @@ This is a changelog for **ZachranObed** application.
 - **ZOB-395** Update meals delivery flow with new design.
 
 ### Removed
+- **ZOB-396** Remove old design components.
 
 ## [1.8.0]
 ### Added